# fix: only get info for parent process