--- a/pkgs/rust-external-deps/default.nix
+++ b/pkgs/rust-external-deps/default.nix
@@ -35,6 +35,7 @@ let
       ]
       ++ lib.optional hostPlatform.isDarwin [
         darwin.libiconv
+        darwin.libobjc
         darwin.apple_sdk.frameworks.SystemConfiguration
       ];