@mkenigs 2024-11-07
> fails to build though
> 
> > ```
> > ld: library not found for -lobjc
> > ```

Had to add
https://github.com/flox/flox/pull/2338/files#diff-e30a8c5c14c33d3f8430ba2c5c42b61d320e9b27001f23d7c146e5bd4c57ca00R38